% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcCoordPath.R
\name{compute_variance_explained}
\alias{compute_variance_explained}
\title{Compute variance explained by latent factors}
\usage{
compute_variance_explained(data_list, coPathRes, view_names = NULL)
}
\arguments{
\item{data_list}{A list of gene expression matrices, one per cell type, with genes in rows and samples in columns.}

\item{coPathRes}{The result object from \code{mcCoordPath}, containing at least the elements \code{Z_list}, \code{B}, and \code{V_list}.}

\item{view_names}{Character vector specifying the names of cell types corresponding to the elements in \code{data_list}.}
}
\value{
A numeric matrix with latent factors in rows and cell types in columns.
Each entry represents the proportion of variance in that cell type’s expression data
explained by the corresponding latent factor.
}
\description{
This function calculates the proportion of variance in each cell type’s gene expression
that is explained by each coordinated (shared) and cell-type–specific latent factor
identified by `mcCoordPath`.
}
