% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcCoordPath_utils.R
\name{crossVal}
\alias{crossVal}
\title{Cross validation of results from mcCoordPath}
\usage{
crossVal(coPathRes, data, pathMat, pathMatcv)
}
\description{
Cross validation of results from mcCoordPath
}
\keyword{internal}
