% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcCoordPath_utils.R
\name{gmt_to_binary_matrix}
\alias{gmt_to_binary_matrix}
\title{generate a binary gene-pathway matrix from a gmt file}
\usage{
gmt_to_binary_matrix(gmt_file)
}
\arguments{
\item{gmt}{file; the gmt file name}
}
\value{
A matrix representing the gene-pathway, row is gene, the column is pathway.
}
\description{
generate a binary gene-pathway matrix from a gmt file
}
