% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcCoordPath.R
\name{infer_ligand_activity}
\alias{infer_ligand_activity}
\title{Infer ligand activity to explain changes in a gene list}
\usage{
infer_ligand_activity(
  genelist,
  ligand_target_matrix,
  ligands,
  background,
  method = c("wilcox", "permutation"),
  n_perm = 1000
)
}
\arguments{
\item{genelist}{Character vector of genes of interest (e.g., differentially expressed genes in receiver cells).}

\item{ligand_target_matrix}{Matrix of predicted transcriptional influences
of ligands on target genes, typically obtained from \pkg{nichenetr}.
Rows represent ligands, columns represent target genes, and values represent
regulatory potential scores.}

\item{ligands}{Character vector of potential ligands to test.}

\item{background}{Character vector of background genes to use as a reference.}

\item{method}{Character string specifying the method to infer ligand activity;
either `"wilcox"` or `"permutation"`. Default = `"wilcox"`.}

\item{nperm}{Integer; number of permutations to perform if `method = "permutation"`.
Default = 1000.}
}
\value{
A data frame (or matrix) containing predicted ligand activity statistics,
  including:
  \itemize{
    \item \code{AUC}: Area under the ROC curve for each ligand.
    \item \code{AUPR}: Area under the precision–recall curve.
    \item \code{mean_diff}: Mean difference in target gene regulatory scores
      between target and background genes.
    \item \code{p_value}: Significance of the activity score (based on Wilcoxon or permutation test).
  }
}
\description{
This function evaluates how well each ligand's predicted target gene profile
 can explain the observed changes in a given set of genes.
It computes ligand activity scores using either a Wilcoxon rank-sum test or
a permutation-based approach.
}
