% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcCoordPath.R
\name{infer_lr_interactions}
\alias{infer_lr_interactions}
\title{Infer ligand–receptor interactions for a given latent factor}
\usage{
infer_lr_interactions(
  coPathRes,
  lf_index = 1,
  topn = 200,
  lr_df = LRpair,
  ligand_target_matrix = NULL,
  method = c("wilcox", "permutation"),
  view_names = NULL
)
}
\arguments{
\item{coPathRes}{A result object from \code{mcCoordPath}}

\item{lf_index}{Integer specifying the index of the latent factor to analyze.}

\item{topn}{Integer; number of top genes contributing to the selected latent factor
to be considered for ligand–receptor interaction analysis.}

\item{lr_df}{Data frame of ligand–receptor pairs, with ligands in the first column
and receptors in the second column.}

\item{ligand_target_matrix}{Optional matrix of predicted transcriptional influences
of ligands on target genes, typically obtained from \pkg{NicheNetR}. Rows represent
ligands, columns represent target genes, and values represent regulatory potential scores.
If \code{NULL}, ligand activity is not predicted.}

\item{method}{Character string specifying the method to infer ligand activity when
\code{ligand_target_matrix} is provided. Either \code{"wilcox"} or \code{"permutation"}.
Default = \code{"wilcox"}.}

\item{view_names}{Character vector specifying the names of cell types corresponding
to the elements in \code{data_list}.}
}
\value{
A data frame (or matrix) containing inferred ligand–receptor interactions, including:
\itemize{
  \item \code{ligand}: Ligand gene name.
  \item \code{receptor}: Receptor gene name.
  \item \code{sender}: Cell type expressing the ligand.
  \item \code{receiver}: Cell type expressing the receptor.
  \item \code{AUC}: Area under the ROC curve for each ligand.
  \item \code{AUPR}: Area under the precision–recall curve.
  \item \code{mean_diff}: Mean difference in regulatory potential between target and background genes.
  \item \code{p_value}: Significance of ligand activity (based on Wilcoxon or permutation test).
}
}
\description{
This function identifies potential ligand–receptor interactions between cell types
that are associated with a specific latent factor from a coordinated multi-view model.
Optionally, it can evaluate ligand activity using NicheNet's ligand–target matrix.
}
