% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcCoordPath_utils.R
\name{solveU}
\alias{solveU}
\title{Compute pathway-LF matrix based on gene-LF and gene-pathway matrices}
\usage{
solveU(
  Z,
  Chat,
  pathMat,
  glm_alpha = 0.9,
  maxPath = 10,
  target.frac = 0.7,
  L3 = NULL
)
}
\arguments{
\item{Z}{matrix; gene-LF matrix with gene in rows and LF in columns.}

\item{Chat}{matrix; regularized inverse of C (gene-pathway matrix), which is used to derive a prior U and select candidate pathways}

\item{pathMat;}{a binary gene-pathway matrix with gene in rows and pathways in columns}

\item{glm_alpha;}{alpha parameter in glmnet}

\item{maxPath;}{the maximum number of pathway to be considered}

\item{target.frac;}{the proportion of latent factors with pathway enriched}

\item{L3;}{the lambada parameter to control the sparsity of pathway}
}
\value{
U matrix; pathway-LF matrix with pathway in rows and LF in columns
}
\description{
Compute pathway-LF matrix based on gene-LF and gene-pathway matrices
}
